--- public/netlify.toml
@@ -1,4 +1,0 @@
-    [[redirects]]
-    from  ="/*"
-    to="index.html"
-    status=200
