+++ src/app/_components/header-notice/_components/notice-item/index.tsx
@@ -1,7 +1,7 @@
-import BusinessTimeIcon from "@/components/common/Icons/business-time-icon";
-import LocateIcon from "@/components/common/Icons/locate-icon";
-import MailIcon from "@/components/common/Icons/mail-icon";
-import PhoneIcon from "@/components/common/Icons/phone-icon";
+import BusinessTimeIcon from "@/components/common/icons/business-time-icon";
+import LocateIcon from "@/components/common/icons/locate-icon";
+import MailIcon from "@/components/common/icons/mail-icon";
+import PhoneIcon from "@/components/common/icons/phone-icon";
 import cn from "@/utils/style/cn";
 
 type noticeItemProps = {
@@ -17,19 +17,15 @@ const IconsMap = {
   businessTime: BusinessTimeIcon,
 };
 
-export default function SidebarNoticeItem({
-  name,
-  href,
-  iconName,
-}: noticeItemProps) {
+export default function NoticeItem({ name, href, iconName }: noticeItemProps) {
   const Icon = IconsMap[iconName];
 
   return (
-    <li className="relative flex flex-nowrap items-center justify-center">
+    <li className="relative flex flex-nowrap items-center justify-center border-b border-solid border-transparent hover:border-header_text">
       <a
         href={href}
         className={cn(
-          "inline-flex border-b border-solid border-transparent text-[13px] font-semibold leading-[1.23] tracking-[0.015em] text-header_text hover:border-header_text",
+          "inline-flex text-[13px] font-semibold leading-[1.23] tracking-[0.015em] text-header_text",
           {
             "pointer-events-none": iconName === "businessTime",
           },
