+++ src/app/_components/header/_components/nav-item/index.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import FoodIcon from "@/components/common/Icons/food-icon";
-import GuideIcon from "@/components/common/Icons/guide-icon";
-import PetIcon from "@/components/common/Icons/pet-icon";
-import ShopIcon from "@/components/common/Icons/shop-icon";
-import SupplyIcon from "@/components/common/Icons/supply-icon";
+import FoodIcon from "@/components/common/icons/food-icon";
+import GuideIcon from "@/components/common/icons/guide-icon";
+import PetIcon from "@/components/common/icons/pet-icon";
+import ShopIcon from "@/components/common/icons/shop-icon";
+import SupplyIcon from "@/components/common/icons/supply-icon";
 import cn from "@/utils/style/cn";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
@@ -33,7 +33,7 @@ export default function NavItem({ name, path, iconName }: navItemProps) {
       <Link
         href={path}
         className={cn(
-          "max-large-screen:!text-[calc(18*100vw/1920)] group relative flex items-end text-base font-semibold uppercase leading-[1.47] hover:!text-header_text_third large-screen:text-[18px]",
+          "group relative flex items-end text-base font-semibold uppercase leading-[1.47] hover:!text-header_text_third max-large-screen:!text-[calc(18*100vw/1920)] large-screen:text-[18px]",
           {
             "after:absolute after:bottom-[-10px] after:left-0 after:h-[1px] after:w-full after:bg-header_text after:content-['']":
               pathName === path,
@@ -53,9 +53,9 @@ export default function NavItem({ name, path, iconName }: navItemProps) {
         {name}
 
         <span
-          className={cn(
-            "absolute bottom-[-10px] left-[50%] h-[1px] w-0 origin-center translate-x-[-50%] bg-header_text transition-all duration-300 group-hover:w-full",
-          )}
+          className={
+            "absolute bottom-[-10px] left-[50%] h-[1px] w-0 origin-center translate-x-[-50%] bg-header_text transition-all duration-300 group-hover:w-full"
+          }
         ></span>
       </Link>
     </li>
