+++ src/app/login/page.tsx
@@ -1,127 +1,11 @@
-import Button from "@/components/common/Button";
+import Button from "@/components/common/button";
 
 export default function Login() {
   return (
     <div className="font-outfit">
       eiyo wtf
       <Button> Test </Button>
-      <Button size="small" variant="third">
-        hello
-      </Button>
-      Lorem ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus
-      magnam dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
-      ipsum dolor sit amet consectetur adipisicing elit. Sit repellendus magnam
-      dolor voluptas sed unde nisi odit atque temporibus officiis magni,
-      assumenda rem rerum, ullam quae similique laborum quis provident! Lorem
+      <Button size="md" variant="Discovery" />
     </div>
   );
 }
