+++ src/components/common/Icons/business-time-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function BusinessTimeIcon(props: SvgProps) {
   return (
