+++ src/components/common/Icons/facebook-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function FacebookIcon(props: SvgProps) {
   return (
