+++ src/components/common/Icons/guide-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function GuideIcon(props: SvgProps) {
   return (
