+++ src/components/common/Icons/instagram-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function InstagramIcon(props: SvgProps) {
   return (
