+++ src/components/common/Icons/locate-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function LocateIcon(props: SvgProps) {
   return (
