+++ src/components/common/Icons/pet-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function PetIcon(props: SvgProps) {
   return (
