+++ src/components/common/Icons/search-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function SearchIcon(props: SvgProps) {
   return (
