+++ src/components/common/Icons/tiktok-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function TiktokIcon(props: SvgProps) {
   return (
