+++ src/components/common/Icons/youtube-icon.tsx
@@ -1,4 +1,4 @@
-import { Svg, SvgProps } from "../Svg";
+import { Svg, SvgProps } from "../svg";
 
 export default function YoutubeIcon(props: SvgProps) {
   return (
